Move DBuffer global layout to layout module
